+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ```
+* Commit: f899594 - v152
+| Author: fjrodafo - Thu, 23 Jan 2025 18:40:35 +0000
+| 
 * Commit: 85de522 - v151
 | Author: fjrodafo - Thu, 23 Jan 2025 15:13:56 +0000
 |